refactor(core): self cli tasks run in same process
