+++ modules/package-manager/src/__tests__/yarn.test.ts
@@ -157,7 +157,11 @@ describe('Yarn', () => {
 		test('filters workspaces by the ones with a version not in the registry', async () => {
 			vi.spyOn(subprocess, 'batch').mockImplementation((calls) => {
 				return Promise.resolve(
-					calls.map(({ args }) => {
+					calls.map((spec) => {
+						if (typeof spec === 'function') {
+							return spec();
+						}
+						const { args } = spec;
 						const versions: Array<string> = [];
 						if (args?.includes('tacos')) {
 							versions.push('1.2.3', '1.2.4');
@@ -181,7 +185,11 @@ describe('Yarn', () => {
 		test('ignores errors', async () => {
 			vi.spyOn(subprocess, 'batch').mockImplementation((calls) => {
 				return Promise.resolve(
-					calls.map(({ args }) => {
+					calls.map((spec) => {
+						if (typeof spec === 'function') {
+							return spec();
+						}
+						const { args } = spec;
 						const versions: Array<string> = [];
 						if (args?.includes('tacos')) {
 							versions.push('1.2.3', '1.2.5');
@@ -205,7 +213,11 @@ describe('Yarn', () => {
 		test('ignores when yarn npm info includes newlines', async () => {
 			vi.spyOn(subprocess, 'batch').mockImplementation((calls) => {
 				return Promise.resolve(
-					calls.map(({ args }) => {
+					calls.map((spec) => {
+						if (typeof spec === 'function') {
+							return spec();
+						}
+						const { args } = spec;
 						const versions: Array<string> = [];
 						if (args?.includes('tacos')) {
 							versions.push('1.2.3', '1.2.5');
@@ -232,7 +244,11 @@ describe('Yarn', () => {
 		test('does not fail for unparseable json', async () => {
 			vi.spyOn(subprocess, 'batch').mockImplementation((calls) => {
 				return Promise.resolve(
-					calls.map(({ args }) => {
+					calls.map((spec) => {
+						if (typeof spec === 'function') {
+							return spec();
+						}
+						const { args } = spec;
 						const versions: Array<string> = [];
 						if (args?.includes('tacos')) {
 							return ['', ''];
