+++ modules/subprocess/src/index.ts
@@ -8,6 +8,8 @@ import type { LogStep } from '@onerepo/logger';
 
 const logger = getLogger();
 
+export type PromiseFn = () => Promise<[string, string]>;
+
 /**
  * The core configuration for {@link run | `run`}, {@link start | `start`}, {@link sudo | `sudo`}, and {@link batch | `batch`} subprocessing.
  * @group Subprocess
@@ -100,7 +102,11 @@ export type RunSpec = {
  * @throws {@link SubprocessError | `SubprocessError`} if not `skipFailures` and the spawned process does not exit cleanly (with code `0`)
  * @see {@link PackageManager.run | `PackageManager.run`} to safely run executables exposed from third party modules.
  */
-export async function run(options: RunSpec): Promise<[string, string]> {
+export async function run(options: RunSpec | PromiseFn): Promise<[string, string]> {
+	if (typeof options === 'function') {
+		return options();
+	}
+
 	return new Promise((resolve, reject) => {
 		const { runDry = false, step: inputStep, ...withoutLogger } = options;
 
@@ -323,7 +329,7 @@ export async function sudo(options: Omit<RunSpec, 'opts'> & { reason?: string })
  * @throws {@link BatchError | `BatchError`} An object that includes a list of all of the {@link SubprocessError | `SubprocessError`}s thrown.
  * @see {@link PackageManager.batch | `PackageManager.batch`} to safely batch executables exposed from third party modules.
  */
-export async function batch(processes: Array<RunSpec>): Promise<Array<[string, string] | Error>> {
+export async function batch(processes: Array<RunSpec | PromiseFn>): Promise<Array<[string, string] | Error>> {
 	const results: Array<[string, string] | Error> = [];
 
 	if (processes.length === 0) {
